process_messages: Improve ordering of actions
This goes through all the new messages and "collects" all of the
actions that were requested first, then performs them afterwards
in the order of accepting invites, updating from wikis, and finally
sleeping. This reduces errors, especially since messages are processed
in the opposite order that they were received in.