User conditions: remove is_shadowbanned
Removes the ability for conditions to be defined based on whether the
author is shadowbanned or not. This is no longer necessary now that
the option to exclude posts from site-wide banned users from the
modqueue exists.

Shadowbanned users' posts can still be set to auto-approve by defining
a condition that matches their username(s) specifically.